fix int32 to int 32

+++ hw05_parallel_execution/run_test.go
@@ -62,7 +62,7 @@ func TestRun(t *testing.T) {
 		err := Run(tasks, workersCount, maxErrorsCount)
 
 		require.Truef(t, errors.Is(err, ErrErrorsLimitExceeded), "actual err - %v", err)
-		require.LessOrEqual(t, int32(maxErrorsCount), int32(errTaskCount), "extra tasks were started")
+		require.LessOrEqual(t, int32(maxErrorsCount), errTaskCount, "extra tasks were started")
 	})
 
 	t.Run("tasks without errors", func(t *testing.T) {
